feat: add new entry points for background and content scripts
Background and content scripts contents are just a dump
message passing demonstration in order to check if they
are working.

Also note that now the name of the output is important because
we reference the name of the file inside manifesto
file. Because of that, now the output filename does
not use the hash and .bundle.js extensions.